+++ pyron-plugin-impl/src/main/scala/com/cloudentity/pyron/plugin/impl/authn/conf.scala
@@ -8,7 +8,10 @@ case class AuthnPluginVerticleConf(
   openApi: Option[AuthnApiOpenApiConf]
 )
 
-case class AuthnMethodConf(tokenHeader: Option[String])
+case class AuthnMethodConf(
+  tokenHeader: Option[String],
+  properties: Option[Map[String,String]]
+)
 
 case class AuthnApiOpenApiConf(
   oauthUrls: Option[OpenApiOauthUrlsConf],